Authorization With Policies(and a admin )

+++ app/Http/Controllers/ThreadController.php
@@ -105,14 +105,10 @@ class ThreadController extends Controller
 		//
 	}
 
-	/**
-	 * Remove the specified resource from storage.
-	 *
-	 * @param  \App\Thread  $thread
-	 * @return \Illuminate\Http\Response
-	 */
 	public function destroy($channel, Thread $thread)
 	{
+		$this->authorize('update', $thread);
+
 		$thread->delete();
 
 		if (request()->wantsJson()) {
